+++ pyprep/prep_pipeline.py
@@ -17,22 +17,20 @@ class PrepPipeline:
     Parameters
     ----------
     raw : mne.raw
-        The data.
+        The data. Channel types must be correctly assigned (e.g.,
+        ocular channels are assigned the type 'eog').
     prep_params : dict
         Parameters of PREP which include at least the following keys:
-
-        - ref_chs : list
-            - A list of channel names to be used for rereferencing
-              [default: all channels]
-
-        - reref_chs : list
+        - ref_chs : list | 'eeg'
+            - A list of channel names to be used for rereferencing.
+               Can be a str 'eeg' to use all EEG channels.
+        - reref_chs : list | 'eeg'
             - A list of channel names to be used for line-noise removed, and
-              referenced [default: all channels]
-
+              referenced. Can be a str 'eeg' to use all EEG channels.
         - line_freqs : array_like
             - list of floats indicating frequencies to be removed.
-              Can be an empty list to skip this step.
-
+              For example for 60Hz you may specify ``np.arange(60, sfreq / 2, 60)``.
+              Specify an empty list to skip the line noise removal step.
     montage : DigMontage
         Digital montage of EEG data.
     ransac : boolean
@@ -42,6 +40,33 @@ class PrepPipeline:
         an int, it will be used as a seed for RandomState.
         If None, the seed will be obtained from the operating system
         (see RandomState for details). Default is None.
+
+    Attributes
+    ----------
+    raw : mne.raw
+        The data including eeg and non eeg channels. It is unprocessed if
+        accessed before the fit method, processed if accessed after a
+        succesful fit method.
+    raw_eeg : mne.raw
+        The only-eeg part of the data. It is unprocessed if accessed before
+        the fit method, processed if accessed after a succesful fit method.
+    raw_non_eeg : mne.raw | None
+        The non-eeg part of the data. It is not processed when calling
+        the fit method. If the input was only EEG it will be None.
+    noisy_channels_original : list
+        bad channels before robust reference
+    bad_before_interpolation : list
+        bad channels after robust reference but before interpolation
+    EEG_before_interpolation : ndarray
+        EEG data in uV before the interpolation
+    reference_before_interpolation : ndarray
+        Reference signal in uV before interpolation.
+    reference_after_interpolation : ndarray
+        Reference signal in uV after interpolation.
+    interpolated_channels : list
+        Names of the interpolated channels.
+    still_noisy_channels : list
+        Names of the noisy channels after interpolation.
 
     References
     ----------
@@ -53,20 +78,50 @@ class PrepPipeline:
 
     def __init__(self, raw, prep_params, montage, ransac=True, random_state=None):
         """Initialize PREP class."""
-        self.raw = raw.copy()
-        self.ch_names = self.raw.ch_names
-        self.raw.set_montage(montage)
-        self.raw.pick_types(eeg=True, eog=False, meg=False)
-        self.ch_names_eeg = self.raw.ch_names
-        self.EEG_raw = self.raw.get_data() * 1e6
+        self.raw_eeg = raw.copy()
+
+        # split eeg and non eeg channels
+        self.ch_names_all = raw.ch_names.copy()
+        self.ch_types_all = raw.get_channel_types()
+        self.ch_names_eeg = [
+            self.ch_names_all[i]
+            for i in range(len(self.ch_names_all))
+            if self.ch_types_all[i] == "eeg"
+        ]
+        self.ch_names_non_eeg = list(set(self.ch_names_all) - set(self.ch_names_eeg))
+        self.raw_eeg.pick_channels(self.ch_names_eeg)
+        if self.ch_names_non_eeg == []:
+            self.raw_non_eeg = None
+        else:
+            self.raw_non_eeg = raw.copy()
+            self.raw_non_eeg.pick_channels(self.ch_names_non_eeg)
+
+        self.raw_eeg.set_montage(montage)
+        # raw_non_eeg may not be compatible with the montage
+        # so it is not set for that object
+
+        self.EEG_raw = self.raw_eeg.get_data() * 1e6
         self.prep_params = prep_params
-        self.sfreq = self.raw.info["sfreq"]
+        if self.prep_params["ref_chs"] == "eeg":
+            self.prep_params["ref_chs"] = self.ch_names_eeg
+        if self.prep_params["reref_chs"] == "eeg":
+            self.prep_params["reref_chs"] = self.ch_names_eeg
+        self.sfreq = self.raw_eeg.info["sfreq"]
         self.ransac = ransac
         self.random_state = check_random_state(random_state)
 
+    @property
+    def raw(self):
+        """Return a version of self.raw_eeg that includes the non-eeg channels."""
+        full_raw = self.raw_eeg.copy()
+        if self.raw_non_eeg is None:
+            return full_raw
+        else:
+            return full_raw.add_channels([self.raw_non_eeg])
+
     def fit(self):
         """Run the whole PREP pipeline."""
-        noisy_detector = NoisyChannels(self.raw, random_state=self.random_state)
+        noisy_detector = NoisyChannels(self.raw_eeg, random_state=self.random_state)
         noisy_detector.find_bad_by_nan_flat()
         # unusable_channels = _union(
         #     noisy_detector.bad_by_nan, noisy_detector.bad_by_flat
@@ -89,17 +144,17 @@ class PrepPipeline:
 
             # Add Trend back
             self.EEG = self.EEG_raw - self.EEG_new + self.EEG_clean
-            self.raw._data = self.EEG * 1e-6
+            self.raw_eeg._data = self.EEG * 1e-6
 
         # Step 3: Referencing
         reference = Reference(
-            self.raw,
+            self.raw_eeg,
             self.prep_params,
             ransac=self.ransac,
             random_state=self.random_state,
         )
         reference.perform_reference()
-        self.raw = reference.raw
+        self.raw_eeg = reference.raw
         self.noisy_channels_original = reference.noisy_channels_original
         self.bad_before_interpolation = reference.bad_before_interpolation
         self.EEG_before_interpolation = reference.EEG_before_interpolation